ragas: bump ragas version, pass old rubric in RubricScore
Before ragas v0.2.11 RubricScores.rubrics wasn't
being applied properly. This commit sets that
as the minimum version for this library.

A change in v0.2.11 from previous versions was a
change in the prompt for domain specific knowledge
evaluation with reference. The prompt from previous
versions is now explicitly passed in.

Signed-off-by: Ali Maredia <amaredia@redhat.com>